@@ -1,15 +1,16 @@
 {
   "name": "@digicatapult/wasp-payload-parser-template",
-  "version": "0.0.4",
+  "version": "1.0.0",
   "description": "Service for WASP",
   "main": "app/index.js",
+  "type": "module",
   "scripts": {
-    "test": "NODE_ENV=test mocha --config ./test/mocharc.js ./test",
+    "test": "NODE_ENV=test mocha --config ./test/mocharc.cjs ./test",
     "lint": "eslint .",
     "depcheck": "depcheck",
     "start": "node app/index.js",
     "dev": "NODE_ENV=dev nodemon app/index.js | pino-colada",
-    "coverage": "nyc npm run test"
+    "coverage": "c8 npm run test"
   },
   "repository": {
     "type": "git",
@@ -28,24 +29,23 @@
   },
   "homepage": "https://github.com/digicatapult/wasp-payload-parser-template#readme",
   "dependencies": {
-    "@digicatapult/wasp-payload-processor": "^1.0.0",
-    "dotenv": "^8.2.0",
-    "envalid": "^7.0.0"
+    "@digicatapult/wasp-payload-processor": "^1.1.43",
+    "dotenv": "^16.3.1",
+    "envalid": "^7.3.1"
   },
   "devDependencies": {
-    "chai": "^4.3.1",
-    "delay": "^5.0.0",
-    "depcheck": "^1.4.0",
-    "eslint": "^7.21.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "kafkajs": "^1.15.0",
-    "mocha": "^8.3.0",
-    "moment": "^2.29.1",
-    "nodemon": "^2.0.7",
-    "nyc": "^15.1.0",
-    "pino-colada": "^2.1.0",
-    "prettier": "^2.2.1",
-    "supertest": "^6.1.3"
+    "chai": "^4.3.7",
+    "depcheck": "^1.4.5",
+    "eslint": "^8.47.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "kafkajs": "^2.2.4",
+    "mocha": "^10.2.0",
+    "moment": "^2.29.4",
+    "nodemon": "^3.0.1",
+    "c8": "^8.0.1",
+    "pino-colada": "^2.2.2",
+    "prettier": "^3.0.2",
+    "supertest": "^6.3.3"
   }
 }
