@@ -1,5 +1,7 @@
-const { buildService } = require('@digicatapult/wasp-payload-processor')
-const { WASP_SENSOR_TYPE } = require('./env')
+import { buildService } from '@digicatapult/wasp-payload-processor'
+import env from './env.js'
+
+const { WASP_SENSOR_TYPE } = env
 
 const payloadProcessor =
   ({ logger }) =>
@@ -42,9 +44,7 @@ const payloadProcessor =
     }
   }
 
-const { startServer, createHttpServer } = buildService({
+export const { startServer, createHttpServer } = await buildService({
   sensorType: WASP_SENSOR_TYPE,
   payloadProcessor,
 })
-
-module.exports = { startServer, createHttpServer }
